refactor: do not pass len in fixed-array encoders
BREAKING CHANGES: Previously, you had to specify the length of the
fixed-array to write.
If the given length was different of the actual array's length,
then an assertion was thrown ("umatched length").

It is no longer possible to specify the length.
As a consequence, the fixed-array writers can no longer assert the
length.

```js
bare.writeU8FixedArray(bc, Uint8Array.of(42, 24), 2) // Previously
bare.writeU8FixedArray(bc, Uint8Array.of(42, 24)) // Now
```