refactor: simplify ByteCursor
ByteCursor is the main abstraction of bare-ts/lib.
A ByteCursor is essentially a view over an area of its buffer.
Bytes can only be read and written in this area.
When this area is not sufficient, a new buffer and a new view
are allocated.
The view may cover a subpart of the buffer, as depicted in the following
figure:

|  view.buffer  |
    | view |

Previously, the entire buffer was copied to a new one.
The reservation failed wheither the new buffer's length overpassed
maxBufferLength or wheither the view did not cover the end of its
buffer, as depicted in the figure.

Now only the area covered by the view is copied to a new buffer.

ByteCursor exposes also a new attribute: `bytes`.
`bytes` is an arrayed view (Uint8Array) over the same buffer as view.
They cover both the same area of the buffer.
This avoids some allocations for copying data and enables users to take
advantage of that.