chore(build): avoid glob ambiguity
`**` is similar to `*` in POSIX shells.
In basha nd zsh `**` recursively include all directories.
To avoid different behaviors, we use `*` instead of `**`.

A better solution could be to use find:

  find src -name '*.ts' -exec esbuild ... {} +

Howver, this looks a bit more complex.
We prefer to restrict our hierarchy of source files to a single
level of subdirectories.