refactor: remove ambient class fields
TypeScript provides "ambient" fields for avoiding to generate
a public ES2022 class field.
TypeScript generates public fields when targeting ES2022
or enabling `useDefineForClassFields`.
esbuild folows TSC when loading `tsconfig.json`.

This seems a good practice to emit public fields for ensuring
that fields are allocated in the order of their declaration.
Order of initialization can safely be re-ordered (may happen because of
control flows).
The only drawback is the lack of support of the feature.
However, by ensuring that `useDefineForClassFields` is turned off
and that we target a version lower than ES2022, this is safe to use.

This change turns ambient fields into public fields and
no longer enables `useDefineForClassFields`.