test: test in dev and non-dev mode
This library provides two modes: dev and default (non-dev) modes.
The dev mdoe is emabeld when the `development` condition is set
or when `NODE_ENV` is set to `development`.

Prerviously, only the dev mode was tested.
We now test both thed ev and non-dev mode.

This allowd to identify some robustness bugs related to non-canonical
encoding. See the `FIXME` comments.